Update README.md
My Favorite Series which is an Anime Series is NAruto Shippuden and Dragonball Super which is one of the most popular series an anime that I watched ever since I was a kid and NAruto is all about Naruto being an outcast and he wants to be acknowledged by everyone and become hokage(Village Leader or President) and the series goes on where he won't give up on being the hokage and achieving his goals and has become acknowledged by everyone. Dragon ball super is all about Goku growing up and committing to get even stronger, his positive mental attitude really shows that he wants to protect everyone in the universe and meet the strongest to go to his limits. Lesson to the both series tells you that being naruto is something that you should be in life, "Never giving up on your dreams and Do the most what you can to be yourself" and Goku tells you that "being strongest doesn't mean you are the bravest but being strong is the means to protect everyone"